remove sqlalchemy models, ditching orm
