+++ server/models/subscription.py
@@ -1,13 +1,13 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
-from sqlalchemy.orm import relationship
-from auth.database import Base
+from pydantic import BaseModel
+from typing import List, Optional
 
-class Subscription(Base):
-    __tablename__ = "subscriptions"
+class SubscriptionCreate(BaseModel):
+    price_id: str
 
-    id = Column(Integer, primary_key=True, index=True)
-    stripe_subscription_id = Column(String, unique=True, index=True)
-    customer_id = Column(Integer, ForeignKey("users.id"))
-    status = Column(String, index=True)
-    
-    customer = relationship("User", back_populates="subscriptions")
+class SubscriptionUpdate(BaseModel):
+    subscription_id: str
+    subscription_item_id: str
+    price_id: str
+
+class SubscriptionCancel(BaseModel):
+    subscription_id: str
