--- server/schemas/auth.py
@@ -1,40 +1,0 @@
-from typing import List, Optional
-from pydantic import BaseModel
-
-class ItemBase(BaseModel):
-    title: str
-    description: Optional[str] = None
-class ItemCreate(ItemBase):
-    pass
-
-class Item(ItemBase):
-    id: int
-    owner_id: int
-    class Config:
-        orm_mode = True
-
-class UserBase(BaseModel):
-    email: str
-
-class UserCreate(UserBase):
-    password: str
-
-class User(UserBase):
-    id: int
-    is_active: bool
-    items: List[Item] = []
-
-    class Config:
-        orm_mode = True
-
-class Token(BaseModel):
-    access_token: str
-
-class Note(BaseModel):
-    id: str
-    title:str
-    description:str
-
-class NoteRequest(BaseModel):
-    title:str
-    description:str
